Time: 1581 ms (5.03%), Space: 117.3 MB (8.82%) - LeetHub

+++ 2221-find-triangular-sum-of-an-array/2221-find-triangular-sum-of-an-array.java
@@ -1,27 +1,27 @@
 class Solution {
     public int triangularSum(int[] nums) {
-        List<Long> coef = getCoef(nums.length);
-        long result = 0;
+        List<Integer> coef = getCoef(nums.length);
+        int result = 0;
         
         for(int i = 0; i < nums.length; i++) {
             result += coef.get(i) * nums[i];
             result %= 10;
         }
         
-        return (int)result;
+        return result;
     }
     
-    List<Long> getCoef(int n) {
-        if (n == 1) return List.of(1l);
+    List<Integer> getCoef(int n) {
+        if (n == 1) return List.of(1);
         
-        List<Long> result = List.of(1l, 1l);
+        List<Integer> result = List.of(1, 1);
         
         for(int i = 2; i < n; i++) {
-            List<Long> next = new ArrayList<>();
+            List<Integer> next = new ArrayList<>();
             
             for(int j = 0; j <= result.size(); j++) {
-                long prev = (j == 0)? 0 : result.get(j - 1);
-                long current = (j == result.size())? 0 : result.get(j);
+                int prev = (j == 0)? 0 : result.get(j - 1);
+                int current = (j == result.size())? 0 : result.get(j);
                 next.add((prev + current) % 10);
             }
